Make profile 'to' configuration override saved 'gitpublishto' value
If the user has the following on .gitpublish:
  [gitpublishprofile "SOMEPROFILE"]
  to = SOMEBODY
and they run:
  $ git-publish --profile SOMEPROFILE --to SOMEBODYELSE
and then run:
  $ git-publish --profile SOMEPROFILE

I believe the intended recipient for the message is SOMEBODY, not
SOMEBODYELSE, as 'to' is explicitly configured in the .gitpublish
profile. But git-publish will send the message to SOMEBODYELSE
instead of SOMEBODY.

This changes the 'gitpublishto' logic so the profile 'to' config
option won't be overridden by 'gitpublishto'.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>